restructured files for future additions
